refactor: EventDto 경로 수정 및 연관된 classpath 재설정
