+++ src/main/java/com/universestay/project/admin/dao/EventDaoImpl.java
@@ -1,11 +1,10 @@
 package com.universestay.project.admin.dao;
 
-import com.universestay.project.dto.EventDto;
+import com.universestay.project.admin.dto.EventDto;
+import java.util.List;
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public class EventDaoImpl implements EventDao {
