+++ src/main/java/com/universestay/project/admin/service/EventService.java
@@ -1,9 +1,7 @@
 package com.universestay.project.admin.service;
 
-import com.universestay.project.dto.EventDto;
-
+import com.universestay.project.admin.dto.EventDto;
 import java.util.List;
-import javax.servlet.http.HttpSession;
 
 public interface EventService {
