+++ src/main/java/com/universestay/project/admin/service/EventServiceImpl.java
@@ -1,12 +1,11 @@
 package com.universestay.project.admin.service;
 
 import com.universestay.project.admin.dao.EventDao;
-import com.universestay.project.dto.EventDto;
+import com.universestay.project.admin.dto.EventDto;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 
 @Service
